[Refactor] display-sub-windows-spells.cpp/h の内容を display-sub-windows.cpp/h へ移動
display-sub-windows-spells.cpp/h の内容は display-sub-windows.cpp/h と分ける
ほどの意義は感じられず、後に他の処理と共通化を行うため display-sub-windows.cpp/h
へ移動する。
