+++ Hengband/Hengband/Hengband.vcxproj.filters
@@ -1529,9 +1529,6 @@
     <ClCompile Include="..\..\src\window\display-sub-windows.cpp">
       <Filter>window</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\window\display-sub-window-spells.cpp">
-      <Filter>window</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\window\main-window-left-frame.cpp">
       <Filter>window</Filter>
     </ClCompile>
@@ -4257,9 +4254,6 @@
       <Filter>window</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\window\display-sub-windows.h">
-      <Filter>window</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\window\display-sub-window-spells.h">
       <Filter>window</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\window\main-window-left-frame.h">
