--- src/window/display-sub-window-spells.h
@@ -1,4 +1,0 @@
-﻿#pragma once
-
-class PlayerType;
-void fix_spell(PlayerType *player_ptr);
