+++ src/window/display-sub-windows.h
@@ -19,6 +19,7 @@ void fix_monster(PlayerType *player_ptr);
 void fix_object(PlayerType *player_ptr);
 void fix_floor_item_list(PlayerType *player_ptr, const int y, const int x);
 void fix_found_item_list(PlayerType *player_ptr);
+void fix_spell(PlayerType *player_ptr);
 void toggle_inventory_equipment(PlayerType *player_ptr);
 
 /*!
